move version out of mechansim definition
